refactor: move helper functions to utils
