+++ src/file.js
@@ -1,11 +1,12 @@
 import { createReadStream, createWriteStream } from 'node:fs';
 import { mkdir, rename, writeFile, unlink } from 'node:fs/promises';
-import { basename, dirname, resolve, isAbsolute } from 'node:path';
+import { basename, dirname, resolve } from 'node:path';
 import { pipeline } from 'node:stream/promises';
+import { resolvePath } from './utils';
 
 const operations = {
   cat: async (currentDir, path) => {
-    const newPath = isAbsolute(path) ? path : resolve(currentDir, path);
+    const newPath = resolvePath(currentDir, path);
     const stream = createReadStream(newPath);
 
     stream.pipe(process.stdout);
@@ -24,14 +25,14 @@ const operations = {
     await mkdir(path);
   },
   rn: async (currentDir, path, newFileName) => {
-    const oldPath = isAbsolute(path) ? path : resolve(currentDir, path);
+    const oldPath = resolvePath(currentDir, path);
     const newPath = resolve(dirname(oldPath), newFileName);
 
     await rename(oldPath, newPath);
   },
   cp: async (currentDir, pathToFile, pathToDirectory) => {
-    const sourcePath = isAbsolute(pathToFile) ? pathToFile : resolve(currentDir, pathToFile);
-    const targetDir = isAbsolute(pathToDirectory) ? pathToDirectory : resolve(currentDir, pathToDirectory);
+    const sourcePath = resolvePath(currentDir, pathToFile);
+    const targetDir = resolvePath(currentDir, pathToDirectory);
     const fileName = basename(sourcePath);
     const destinationPath = resolve(targetDir, fileName);
 
@@ -40,8 +41,8 @@ const operations = {
     await pipeline(readStream, writeStream);
   },
   mv: async (currentDir, pathToFile, pathToDirectory) => {
-    const sourcePath = isAbsolute(pathToFile) ? pathToFile : resolve(currentDir, pathToFile);
-    const targetDir = isAbsolute(pathToDirectory) ? pathToDirectory : resolve(currentDir, pathToDirectory);
+    const sourcePath = resolvePath(currentDir, pathToFile);
+    const targetDir = resolvePath(currentDir, pathToDirectory);
     const fileName = basename(sourcePath);
     const destinationPath = resolve(targetDir, fileName);
 
@@ -51,7 +52,7 @@ const operations = {
     await unlink(sourcePath);
   },
   rm: async (currentDir, path) => {
-    const newPath = isAbsolute(path) ? path : resolve(currentDir, path);
+    const newPath = resolvePath(currentDir, path);
 
     await unlink(newPath);
   }
